Release SRDEquote object during PossumPipe reset.
Testing with repetitive connections indicated that the support in
the SRDEquote object for certificate processing was resulting in
an object failure after the object was used.  The SRDEquote
object could arguably use a ->reset() method in order to allow
its repetitive use.

In the interests of time constraints the situation was addressed
by calling a destructor on the object which forces its
re-initialization for each connection.