Isolate credential provisioning into separate function.
This update adds the _process_request() function to the
Provisioner enclave.  The actual processing of a provisioning
request after the security context has been negotiated was moved
into this function.

The primary connection processing loop was also modified to not
generate an exit condition if an error is encountered while
setting up the security context via the
PossumPipe->start_host_mode2 method.  This makes the provisioner
robust with respect to halting if invalid credentials or an error
on the client side is experienced.