Implement support for Provisioner client verification.
This commit implements support for verifying the status of the
Attestation enclave that is requesting provisioning of
credentials.  Currently the decision is made on the following
characteristics:

Enclave attributes.

IDfusion debug signing key.

Vendor ID.

Vendor SVN.

In order to support this initial implementation of enclave
verification the following structure was defined in the SRDE.h
header file:

SRDEendpoint

This structure allows the specification of the following enclave
identity characteristics:

Enclave attributes.

Enclave software vendor identification.

Enclave software security version.

Enclave signer.

Enclave measurement.

The SRDEenpoint structure also contains an element named mask
that implements a bitfield to indicate which of these
characteristics should be used for making the endpoint decision.
This allows very precise control to be implement with a common
data structure.

Given the generic need for endpoint verification the plan will be
to implement all of the needed functionality into an object.