Implement anonymized client identities.
In order to determine how many unique clients are provisioning
credentials this update implements a strategy of using a
statically generated key to identify each client that provisions
credentials.

The provisioner implements a random 256 bit number that is
provided to each client at connection time.  This number is used
as the key identifier (keyid) for generating an enclave static
key.

The keys that ENCLU[EGETKEY] generates are 128 bits in length.
The key is hashed with SHA256 to generate the identity that is
returned to the Provisioner enclave.

This update also addresses a regression which was causing only
the first provisioning event to provision legitimate keys.  This
was caused by the _process_request() function in the Provisioner
object directly using the Buffer objects that were holding the SPID
and APIkey values directly in the PossumPipe->send_packet method.

The arguement Buffer to the ->send_packet method gets modified by
the method call thus corrupting its value.  This was addressed by
using a functional local Buffer object to send the values to the
client.