Add support for identifying connection client.
Support was needed in the PossumPipe client to identify the
identity of the client that was negotiating a security context.
The following method was implemented to surface this information:

PossumPipe->get_client

Identification of the client is currently supported only for mode
2 connections.  In this case the ->get_client method returns a
pointer to the RSAkey object that authenticated the client
connection.