Implement client identification in Provisioner.
In order to track which client is provisioning credentials this
update implements the generation of a unique and anonymous
identity for each client that requests a provisioning event.  The
identity is the SHA256 signature of the modulus of the public key
that is used to authenticate the connection to the provisioning
service.

The ECALL that registers client keys prints out the key signature
when the key is loaded.  This provides a reference for the
provisioning events that output the client identity on each
provisioning event.