Implement explicit PossumPipe object initialization.
Multiple connection testing with lost object support indicated
that the PossumPipe object was losing objects when multiple
security contexts were being created and released during a single
instance of a PossumPipe object.

This was traced to the fact that the PossumPipe object was
allocating state objects at the time a security context was being
negotiated.  This caused the objects to be lost when a new
security context was created.

This was fixed by moving to an explicit object allocation model.
The constructor for the PossumPipe object now creates the
instances of the internal state objects when the PossumPipe
object is created.  The exception to this is the SRDEquote object
that does not yet have a ->reset method available.