Implement function to verify SRDEpipe endpoints.
This commit adds the following method to the SRDEpipe object:

->verify

This method is used to verify the identity characteristics of the
enclave on the other side of the communications conduit.

The method is passed a Buffer object that is used to hold the
array of SRDEendpoint structures that will be used in the
evaluation process.  The first structure that matches the
identity characteristics of the counter-party enclave will
trigger a successful return from the the method.

The success of evaluating the endpoint is specified via a boolean
variable accessed via a pointer passed to the method.  A true
value indicates that a match was encountered on one of the access
control structures.