Implement endpoint verification for attestation services.
This update uses the SRDEpipe endpoint verification
infrastructure to implement endpoint verification for attestation
services.

With respect to the attestation enclave any initiating enclave is
allowed to connect with the only requirement being the enclave is
initialized and in either debug or production mode.

A precise definition for the Provisioner enclave is used except
for the measurement of the enclave.

Precise definitions with the exception of the measurement value
are used in the LocalTarget enclave for defining the connection
to the attestation services enclave.