Fix definition of ->verify mode in SRDEpipe definition.
A warning message had been missed with respect to the fact that
in the standard userspace definition of the ->verify method a
boolean value was being used rather then a pointer to a boolean.

There was no operational effect to this since the standard
userspace implementation does not use the ->verify method.