Implement canonical enclave definitions.
The ENCLAVE_DIR definition when passed to the build system had
previously triggered the specification of the pathname to an
enclave being loaded.  Since it was more or less a canonical
indication that this was a production build the CPP instance of
the definition in source files had been used to trigger certain
actions.

Since there is now a need to specify whether enclaves should be
loaded in production or debug mode the decision was made to
implement a make variable that would be used to specify that this
was a production build.  The following variable was chosen:

SRDE_PRODUCTION

Which if defined will be interpreted as a signal that compilation
should be configured in a way tha would assume that all paths and
locations be rooted in the following two directory locations:

/opt/IDfusion

/var/lib/IDfusion

The following two CPP macros were defined to implement setting
the location of an enclave and its initialization token:

ENCLAVE_LOCN

TOKEN_LOCN

The ENCLAVE_LOCN macro is give the filename of the enclave, for
example enclave.signed.so and places it in either the current
working directory or in /opt/IDfusion/lib/enclaves.

The TOKEN_LOCN macro is given the filename of the initialization
token and places it either in the current working directory or in
/var/lib/IDfusion/tokens.

The following CPP macro is used to provide an indication of
whether or not the enclave should be run in debug mode or
non-debug mode.  As one would anticipate a build with
SRDE_PRODUCTION=true will signal that the enclave is to be loaded
in non-debug mode.

In the previous strategy the ENCLAVE_NAME CPP define would
translate into the name of the enclave.  Until the source can be
completely audited to remove the use of this define it was
formulated to use the ENCLAVE_LOCN macro.