Fix PossumPipe and SRDEpipe ->verify method.
Functional testing with non-debug enclaves indicated that the
->verify method in both the PossumPipe and SRDEpipe objects were
only using the first entry in an SRDEendpoint array to evaluate
the identity characteristics of the remote entity.  This was
secondary to the fact that the endpoing processing loop had
evolved from a for loop to a while loop.

The previous implementation with a for loop was incrementing the
pointer used to point to a specific structure within the endpoint
array list.  When the loop was converted to a while loop the
increment of that pointer had not been moved into the loop body.

This commit adds the missing pointer increment in both object.