Fix seal key generation.
Testing with production enclaves had indicated what appears to be
a significant anomaly in the ability of the SEALkey object to
generate sealing keys.  The SEALkey object was returning an error
code secondary to the following conditional:

if ( !(report.body.attributes.flags & 0x0000000000000001ULL) || \
     !(report.body.attributes.flags & 0x0000000000000002ULL) )
		ERR(goto done);

This check had been followed from the Intel SDK sources in order
to verify the ability to migrate seal keys from one enclave to
another.

On its face the conditional rejects generation of a key for an
enclave that reports that the INIT attribute bit or the DEBUG
attribute bit are not set.  Running an enclave in production mode
without the debug bit set obviously causes the conditional to
fail.

It is completely unclear why the conditional should be
constructed in this manner.  It would possibly make sense if the
logical negation were left out in the second clause.  This would
cause enclaves running in debug mode to reject key generation
which would arguably be a security feature.

This commit drops the second conditional completely in order to
create operational production enclaves.