Fix attestation services endpoint descriptions.
The Provisioner enclave was missing a definition for production
enclaves, a situation that caused the Provisioner to fail
connections when the production implementation was tested.  This
commit provides a clause to define production enclaves.

In the case of both the Provisioner and Attestation enclaves the
debug endpoint definition for provisioning was dropped based on a
test for the SRDE_PRODUCTION define.  This insures full
protection of the confidentiality of the attestation keys.