Add LocalTarget.signed.so as a production enclave.
Due to the build procedures that are now being embraced with the
SRDE_PRODUCTION define the LocalTarget.signed.so enclave was
configured in the provision-tokens utility to have its token
generated with debug disabled.

The srde-attestation utilility running in test mode was failing
since it was configuring the LocalTarget enclave to run in
non-debug mode.