Fix unit test sample Makefile.
Deployment testing indicated that the sample unit tests were
failing secondary to two reasons.

Due to the introduction of dynamic OCALL management, compilation
of the enclave requires access to the SRDEfusion-ocall.h and
SRDEnaaam-ocall.h include files that are in
/opt/IDfusion/include/SRDEtime.  An include directive was added
to the enclave CFLAGS in order to include this directory in the
default search path.

Removal of the dependencies on the enclave implementation of the
OpenSSL library resulted in duplicate symbol definitions for the
_strdup function.  The enclave linker directive was modified to
drop linkage against the the enclave C++ library.

This latter issue may be problematic for building libraries that
have C++ dependencies.  Consideration may have to be given to the
use of a weak symbol directive for this function.