Implement Attestation object.
Since it is anticipated that multiple enclaves will want to take
advantage of attestation services an Attestation object was added
to the enclave NAAAIM library.  This object encapsulates the
functionality that had previously been in the LocalTarget ECALL
that tested attestation services.

The LocalTarget enclave was updated to use an Attestation object
rather then open coding of the interface.

Useability refinements will be pending for the Attestation
enclave.  It currently demonstrates the ability to generate a
simple attestation report for an enclave.