Fix issues with reduced length IAS nonces.
A regression was noted with IAS attestation reports that were
being requested by enclave versions of the SRDEquote object.  If
an IAS nonce less then 16 bytes in length was being requested
information from enclave memory would be leaked in the remaining
bytes of the nonce.

This was secondary to the OCALL code using the array size of the
nonce in the OCALL marshalling object rather then the actual
length of the supplied nonce.

There was a boolean variable in the OCALL marshaling structure
that was being used to indicate to the SRDEquote manager that an
IAS nonce was to be used.  This variable was dropped and replaced
with a variable that indicates the size of the nonce that will be
used.  The SRDEquote manager uses a non-zero value of this
variable to indicate that a nonce should be processed.