Implement IAS report nonces and data for Attestation object.
The initial implementation of the attestation object did not
support the ability to specify either an IAS report nonce or
reportdata that is to be authenticated by the attestation report.
This commit provides that infrastructure.

Until ASN.1 encoding of the report request can be implemented a
report nonce is requested by appending up to 16 bytes of nonce
data to the packet that returns the enclave report that is to be
signed by the quoting enclave.  The Attestation object verifies
that any supplied data does not exceed 16 bytes in length.

The LocalTarget enclave unit test for the Attestation enclave was
updated to generate a random nonce and random report data to
verify the veracity of this functionality.