Implement method to surface report elements.
Both the standard and enclave versions of the SRDEquote object
were updated to include a ->get_report method in order to allow
the three principal elements of an Intel Attestation Services
(IAS) to be surfaced to application code.

These three elements are as follows:

Report certificates.

Report signature.

JSON encoded report.

The test-remote-target utility was modified to implement a -f
command-line switch that selects for the previous behavior where
a full characterization of the IAS report is output.

Absent this command-line arguement an abbreviated output is
generated where only the above three elements are output.