Update Changes to document Intel development libraries.
Prior to tagging of the version 0.9 release a review was
conducted of the Changes file.  A decision was made to add an
item that documented that release 0.9 is using a 2.5.5 variant of
the Intel development libraries.

While release 0.9 is based on the version 2.6 pre-built enclaves
a regression was encoutered with the full release of the 2.6
development libraries.  To address this issue the full 2.6
library release was used with the libsgx_tstdc.a from the version
2.5 release that did not demonstrate the regression.