Fix multiple mode2 PossumPipe clients.
Field experierence with provisioning attestation service
credentials indicated that provisioning failures were occurring
for any client identities other then the first one that is loaded
by the srde-provisioner utility.

On the client side a connection would be made that would just
hang.  On the server side the following log messages were
generated:

Waiting for connection.
Have connection, setting up context.
T[../../../lib/SHA256_hmac.c,add_Buffer,279]: Error location.
T[PossumPipe.c,find_client2,1141]: Error location.
        Unable to start connection.

Based on the stack trace the problem was quickly located.  The
loop which was traversing over the client identities was not
resetting the SHA256_hmac object that was being used to compute
the client identity.

The SHA256_hmac object contains an internal check that causes the
object to be poisoned if the ->add_Buffer method is called after
the ->compute method has been invoked without an intermediate
call to the ->clear method.

The resolution of the problem was thus straight forward.  A call
to the ->reset method of the SHA256_hmac object that was
computing the client identity was added at the end of the loop in
order to prepare the object for the next client iteration.

One of the issues that still needs to be addressed is why the
client hangs rather then issueing some type of informative error
message.  Addressing this will be the subject of a subsequent
commit.