Implemented automated path selection for attestation services.
This commit uses the recently introduced ENCLAVE_LOCN and
TOKEN_LOCN macros to automatically select the location of the
enclave that is to be loaded and the location of the launch token
for that enclave.

In addition the ENCLAVE_DEBUG macro is used to define whether or
not the enclave is to be initialized in debug or production mode.

The definition for the pathname of the file that contains the
sealed attestation credentials was place inside of clauses that
are gated by the SRDE_PRODUCTION macro.  If this macro value
evaluates to true the following standard location is used:

/var/lib/IDfusion/data/Attestation.bin

In the event of a non-production build the path definition is
removed and the current working directory is used as the source
of the Attestation.bin file.

These changes allow development versions of attestation services
to be tested without disturbing a production instance
installation.