Detect EOF on enclave Duct reads.
A regression in PossumPipe was found where an error on the server
side was not causing the client to detect that the connection was
closed by the server.

This was finally linked to the fact that the enclave
implementation of the Duct objecting was not detecting an end of
connection indication at the Duct protocol level.  The protocol
that Duct uses interprets a zero length write as an indication
that the writing party is closing the connection.

The enclave implementation of the Duct object had the ability to
detect an EOF condition in the standard userspace implementation
of the object but this requires an independent OCALL when the
information is already available as the result of a read.

This update modifies the ->receive_Buffer method to check for a
zero length receive after a check is made that the receive had
successfully completed.  The EOF flag in the state object is then
set.