Signal end of connection on failed client lookup.
A regression had been noted in attestation credential
provisioning that was causing the client to simply hang in
FIN_WAIT2 state if the provisioning server was unable to find a
matching client in its identity registry.

This commit addresses this problem by sending a zero length
message if a matching client identity cannot be found.  This
triggers end of connection detection by the enclave
implementation of the Duct object causing the client to properly
terminate.