Output reason for connection failure.
With proper detection of an invalid client identity in place the
error state of the pipe was interrogated after the
->start_host_mode2 method was invoked in order to find the error
code associated with the failure to setup the security context.

If the error is secondary to an invalid client an appropriate
message is output, otherwise a general connection failure message
is output.