Implement unified detection of production builds.
The SRDE_PRODUCTION make variable is used to indicate whether or
not the build is a production or debug implementation.  This make
variable needs to be extended into both standard and enclave
compilations as a CPP variable in order to trigger the build
changes.

This commit implements a unified framework for doing this.  A
single test is made at the top of the makefile and used to set
the TYPE make variable.  This variable is then included in the
CFLAGS and ENCLAVE_CFLAGS variables as a command line CPP macro
definition.