Document SDK and runtime update.
The Changes documentation was updated to reflect that as of this
point the SRDE has been documented to build and run with version
2.7.1 of the Intel SDK libraries and prebuilt enclaves.

This update includes the version of the Intel signing key white
list that was active as of 1/7/2020.

This update is significant in the fact that the previous release
had been using version 2.6 libraries with a version of the
libsgx_tstdc.a library from version 2.5.  This was secondary to
the signing tool failing when used with enclaves that were built
with versions of the library after 2.5.

The signing was failing secondary to the fact that the signing
utility was detecting the presence of a .ctors or construction
section.  Since such a section interferes with global
constructor execution for the enclave the signing exits with an
error condition.

The source of the constructor sections was the libirc library that
are part of the pre-built libraries downloaded from Intel and
incorporated into the libsgx_tstdc.a library.  The libirc library
are a series of optimized C routines from the Intel compiler
suite and are not available in source code form.

The root cause of this problem was found to be versions of the
GNU binutils that do not properly elide constructor sections.
Testing suggests that the problem is present in version 2.26 of
the binutils library but is not present in the 2.27 release.