Implement initial framework for SRDE signing system.
This update provides the initial step forward in implementing an
SRDE specific enclaving signing system.  At the current time
the actual signing of enclaves are still dependent on the Intel
signing utility.

An SRDE specific signing utility will ultimately be named as
follows:

srde-sign

This commit provides the implementation of a shell script by that
name that wraps execution of the Intel signing binary.

The primary motivation for this initial support is to provide a
mechanism for automating the generation of production enclaves.
Previous to this commit the individual enclave metadata
configuration files in XML format needed to be edited in order to
disable the Debug attribute bit so that production enclaves had
to be initialized in non-debug mode.

The srde-sign utility uses a new form of metadata configuration
file which has one configuration item per line with the name of
the configuration item separated from its value by a space.  The
utility converts this form into an XML file suitable for use with
the sgx_sign utility.

The srde-sign utility takes as a mode arguement the -P switch
that causes the utility to generate an XML configuration file
with the debug disable directie.  By default, without the -P
arguement, the srde-sign utility adds a directive to not disable
debug mode.

If the SRDE_PRODUCTION make variable is defined the -P
command-line option is automatically added to the srde-sign
utility in order to cause any enclaves signed by the utility to
have debug mode disabled.

Currently only the enclaves in the SRDE/Attestation directive
have been converted to the srde-sign utility.  These enclaves are
the ones that currently need to be configured to disable debug
mode.

A new make variable named INSTALLDEV was added to the
SRDE/Makefile that defines binaries that are to be installed as a
component of the install-dev target.