Remove unneeded attribute check.
In the following commit:

commit 508d34ddbd9210b6fb6b6c4bf54109dc
"Fix seal key generation."

A portion of the clause that checks the attributes returned by
the self report that provides the parameters for the seal key
generation was removed since it was causing seal key generation
by production key enclaves to fail.  At the time of the commit it
was documented that this may be a failing in the Intel SDK.

The generation of seal keys was tested using the Intel functions
for generating seal keys and those routines worked as expected.

Further analysis indicated that the attribute check in our code
was not needed given that our implementation uses a procedure
different from the Intel code.  The Intel code sets the security
relevant attributes and then calls the sgx_seal_data_ex()
function that generates the sealing key.

The attribute check in the latter function, in retrospect, seems
to have been implemented secondary to a 'belts and suspender'
approach to make sure that the attributes are in a minimally
secure posture.  In our implementation we explicitly set
attributes that are correct from a security perspective so our
attribute check is irrelevant since it is acting on the
attributes returned by the enclave self-report.