Remove branding from GUI clients.
This update removes braning from the graphical user interface
clients.  The idclient utility which is not currently used now
uses a generic title.

The AImanager utility was converted to use a generic name for the
logo to be used.  The previous logo was not entrained in the
source tree so the actual logo to be used will need to be
addressed.