Convert to generic key naming nomenclature.
The production and debug keys for enclaves are designed to be
embedded in the source files for a given entity.  Rather then
using a branding specific variable name this update converts to
the following names:

Debug_key

Production_key

This commit also converts the key definition to be all nulls.
Given the this code is designed to be generic a scheme needs to
be developed for automatically generating the signature of the
signing keys.