Implement generic program copyright banner.
This update tests a generic implementation for copyright
bannering for applications.  A general copyright banner of the
following form is implemented:

"%s: Copyright (c) %s, %s. All rights reserved.\n"

This is designed to have the string fields replaced with the
following, in order:

Program name

Date of copyright

Company declaring copyright

While this update has the date and the company declaration
localized to this file the plans will be to lift them into a
generic header file that will allow a global update of this
information.