Update kernel constants for 5.4 kernel implementation.
This update modifies the kernel system call number for modifying
the security state of a process and the value for creating a new
security events namespace.  These values are consistent with port
of even modeling to the 5.4 kernel.

The numeric value for the namespace definition have been moved
into the low order bits of the namespace mask since all of the
namespace bit possitions have been consumed in the 5.4 kernel.
This kernel has the namespace bits expanded to 64 bits but the
kref structure still uses an int value to hold the namespace
identifier.

The bit value for the event namespace was chosen to avoid
conflict with the CLONE_NEWTIME namespace which also uses one of
lower bits (0x80) in the clone flags.  These lower order bits can
only be used with the clone3 system call or as an argument to
the unshare system call.