Add support for selecting an external authenticator.
The -e command-line switch to the cboot utility was previously
used to specify the name of the enclave to be used when SGX based
event evaluation was being used.  SGX is no longer a primary
evaluator and previous development has suggested there is limited
utility to expending a command-line switch to specify an enclave.

As a result the -e command-line switch was re-appropriated to be
used to specify that an event namespace is to use external
evaluation.  A boolean variable is set indicating this selection
that is passed to the setup_namespace() function.

When this variable is detected to be true the function writes an
ASCII 1 to the following pseudo-file:

/sys/kernel/security/integrity/events/external

After the event namespace is unshared.  This places all
responsibility for maintaining the event domain in whatever
external evaluator is being run by the cboot utility.