Implement explicit sleep/wakeup processing.
This update implements support for kernel event modeling that
places tasks in interruptible sleep until the external evaluator
makes a decision as to the modeling mode and the trust state of
an event.

If the event namespace is sealed against modifications an event
must correspond to the set of trusted events.  If it does the
sys_bad_actor system call is used to wake the process up and
clear the bad_actor status bit in the task control structure.

If the event is not specified as trusted the process is woken up
but the bad_actor status bit remains set.  This causes the TE lsm
to discipline the process.

If the event namespace is 'open', ie. not in a sealed and
unmodifiable state the sys_bad_actor system call is issued to
clear the bad_actor status of the bit and wake the process up.

This model implements a 'default secure' model since the external
evaluator needs to make an explicit decision on the trust state
of the context of execution in order to allow the process to run
in an undisciplined state.