Stash event domain aggregate value.
In preparation for adding support for computing the state
measurement of an event domain the domain model implementation
was modified to save the host extended aggregate value in the
state of the model.  This value is the first measurement value
that is added to a domain.

This stashed value will be used as the measurement starting point
for the domain state value and will avoid having to compute the
value needlessly each time the state value is requested.