Generalize extension summing.
In order to support the use of the code that updates the
measurement of a domain to be applicable to the generation of the
state value the _update_measurement private function was renamed
to _extend_measurement with functional changes.

The function was modified to take an arguement that specifies the
output of the measurement value.  The call sites were updated to
reflect this change.

This will allow the extension code to be used to compute the
state measurement over the sorted list of event points.