Implement support for event domain state computation.
The mathematical 'state' of an event domain is defined as the
extension sum of the domain event points that have been sorted in
natural hash (big-endian) format.  The root of the extension sum
is the domain extended aggregate measurement of the platform.

This update provides support for computation of the state value
in the domain event model object.

The domain boot and management utilities were updated to support
output of the state value for a domain with the following command
in the management utility:

show state

The event domain test harness was updated to output the state
value by specifying the -M command-line arguement to the utility.