Drop CAP_TRUST before launching event domain.
With kernel implementation of CAP_TRUST the cboot utility was
modified to drop the CAP_TRUST capability from the bounding set
of capabilities after unsharing the event domain namespace.

This is a first step in the path toward having only a single
supervisory entity have control over the trust state of the
platform.