Update tests.
Issues were noted with the build of the test files.  This commit
addresses these issues in preparation for adding the TTYduct object that
will be used by the new Quixote infrastructure.

The build of the OTEDKS_test utility is being temporarily dropped by
this commit since it needs more generic updates.