Add TTYduct object.
This update implements an object for conducting packet based
data transfers over a tty object.  The primary intent of this object is
to provide similar functionality to that provied by the Duct and
LocalDuct objects.

The TTYduct object is still very much preliminary and has a large cohort
of code that is selected by #if statements at the current time.  This is
to support debugging of object functionality against a micro-controller
based Sancho security co-processor.