Initial implementation of security state manager.
This commit introduces Quixote which will be the new utility for
running a process in an independent security state namespace.

The objective of Quixote will be to have loadable modules that
implement the various roots of trust in which the the security
state modeling engine will be run.  The current implementation is
specific to a micro-controller based Sancho implementation.