Add directory framework for Sancho modules.
This update implements the directory that will contain a heirarchy for
building the Sancho modules that provide the security state modeling
engines.

This initial implementation contains a framework for building the Sancho
security co-processors based on the STmicro 32 bit ARM controllers.

The initial implementation framework build an ARM specific
implementation of the components from the HurdLib support library.