Implement build infrastructure for cellular enabled Sancho.
This update introduces the directory build heirarchy for the ST
micro-controller implementation of a Sancho security co-processor that
is cellular enabled.

The heirarchy includes the build of all of the support libraries that
the primary application depends on.

The strategy behind this heirarchy is to use target directives that
specify the directory in the STmicro supplied source code from which to
abstract the source files.