Unify ST micro-controller build system.
This update implements a standard source for build directives for ST
micro-controllers.  The strategy is to place all generic build
directives for the ST micro-controllers in the following file:

Sancho/STmcu/STmcu.mk

The subordinates makefiles are designed to sense whether or not the
build is being generated in the current directory or whether it is a
part of a recursive build.  In the former case, where make is executed
in a specific sub-directory, the standard MCU directive file is
referenced directly.

Significant more work needs to be done to rationalize the build process
for the micro-controllers.