Add ST micro-controller Sancho implementation.
This commit introduces a preliminary implementation of a Sancho instance
based on the stm32l4xx micro-controller.  This instance includes support
for external reporting via a cellular modem connection.

The current implementation is hanging when an exchange event is added
immediately after a seal event.

In addition there are signficiant other issues to be addressed which
will now be tracked as updates to this source base.