Remove use of packed structure for security state values.
The original implementation of the security state modeling code depended
on the use of a packed structure to hold the elements that define the
security state for a context of execution or an execution target.  This
turns out to be problematic with respect to modeling compatibility
between different platforms, most notably x86 and ARM.

The modeling code dereferences pointers in the structures which can
result in an unaligned pointer dereference in a packed structure on
the 32-bit ARM micro-controller implementations of Sancho.

This update drops the use of the packed structure declaration.  Instead
the code to compute the entity specific contribution to a security state
incrementally adds the structure elements to the buffer that will be
submitted to the SHA256 hashing function.  This implements the same byte
sequence feed to the hash function with a universal implementation that
will work regardless of structure element alignment.

The new code was tested to verify that it yeilds the same security state
and measurement values as the code that uses a packed structure.