Explicitly cast arguments to state creation snprintf call.
Issues were noted with respect to the GCC compiler being used to compile
the x86_64 implementation of the Actor and Subject objects and the
compiler used for the ARM Sancho implementation.  The call to the
snprintf function that is used to create the ASCII description of a
state component was generating format warnings either with the GCC or
ARM compiler depending on the format specifiers that were being used.

In order to allow a single source file for both implementations explicit
casts to the arguments were added that pleased both compilers.  The
resulting code generates equivalent measurement and trajectory
descriptions on both platforms with this fix.

It may be possible to re-visit this issue when the compiler used for the
x86_64 code is upgraded.