Remove event time accounting.
The initial implementation of the quixote cartridge manager had an
attempt at determining the amount of time that was spent by the Sancho
implementation in processing a security state event.  That code was
accounting process time rather then wall time, with the former not being
of as much interest at the latter given the desire to determine the
latency induced by the modeling implementation.

This commit drops that implementation that will be replaced at a later
time by an implementation that uses wall time for the measurement.