Implement state display by Sancho MCU implementation.
The newer implementation of the security state modeling code implements
support for the computation of the state of a security domain which
represents a time invariant measurement of the status of the domain.

This commit adds support to the quixote cartridge manager and the MCU
implementation to return that value.