Fix MCU response handling.
A significant regression was noted in MCU modeling where communications
between the cartridge manager and the MCU implementation would hang
after a 'seal' event was given to the MCU implementation.

After a significant amount of debugging effort this was found to be
secondary to XON processing having been inadvertently enabled in the
TTYduct object secondary to an attempt to debug another connection
issues.  This was resulting in any binary value of 17 (0x11) being
intercepted by the TTY driver and not reaching the TTYduct object.

The command string to invoke discipling of a process that has a five
digit PID number is 17 characters in length, including the NULL string
terminator, ie:

DISCIPLINE NNNNN

This caused the TTYduct object on the MCU side to send the following
packet length descriptor, in hexadecimal big endian format:

00 00 00 11

The hexadecimal value of 11 (17 decimal) was being consumed by the TTY
layer which caused the first byte of the actual message to be
interpreted as the packet length.  Since the MCU was not sending a
packet of that length quixote would hang waiting for what it thought was
an additional portion of the message.

This commit also changes the _read_buffer implementation to use polling
along with FIONREAD processing which is more efficient then the single
character implementation that was previously used.  This latter
implementation was implemented in order to debug the XON hand processing
previouisly described.